requirements.txt
torch
transformers
numpy
faiss-cpu
fastapi
uvicorn